3380.2.py
In this version of the python script I've decided to add a menu. Rather than just pulling a random achievement I've created a interactive menu for you to choose a game based on its steam ID, pull a random achievement, pull all of the games achievements or all of your personal unlocked achievements for a game.